# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Talent Graph API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new referrer account (clients join by admin invite only)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterInput" }
      responses:
        "201":
          description: Created account
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthUser" }
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Log in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginInput" }
      responses:
        "200":
          description: Authenticated user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthUser" }
...
  /auth/verify-email:
    post:
      operationId: verifyEmail
      tags: [auth]
      summary: Verify an email address with a token from the verification email
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VerifyEmailInput" }
      responses:
        "200":
          description: Authenticated, verified user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthUser" }
...
  /auth/request-password-reset:
    post:
      operationId: requestPasswordReset
      tags: [auth]
      summary: Request a password reset email
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RequestPasswordResetInput" }
      responses:
        "200":
          description: Reset email sent if the account exists
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageResponse" }
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Set a new password using a reset token
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ResetPasswordInput" }
      responses:
        "200":
          description: Password updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageResponse" }
  /auth/invite/{token}:
    get:
      operationId: getInvite
      tags: [auth]
      summary: Preview an invite (email, name, company) before accepting
      parameters:
        - name: token
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Invite details
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvitePreview" }
  /auth/accept-invite:
    post:
      operationId: acceptInvite
      tags: [auth]
      summary: Accept a client invite by setting a password
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AcceptInviteInput" }
      responses:
        "200":
          description: Authenticated, activated user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthUser" }
...
components:
  schemas:
...
    AuthUser:
      type: object
      required: [id, name, email, role, emailVerified]
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string }
        role: { type: string, enum: [ADMIN, CLIENT, REFERRER] }
        emailVerified: { type: boolean }
        referrerCandidateId: { type: ["string", "null"] }
        clientId: { type: ["string", "null"] }
        clientName: { type: ["string", "null"] }
    MessageResponse:
      type: object
      required: [message]
      properties:
        message: { type: string }
    VerifyEmailInput:
      type: object
      required: [token]
      properties:
        token: { type: string }
    RequestPasswordResetInput:
      type: object
      required: [email]
      properties:
        email: { type: string }
    ResetPasswordInput:
      type: object
      required: [token, password]
      properties:
        token: { type: string }
        password: { type: string }
    AcceptInviteInput:
      type: object
      required: [token, password]
      properties:
        token: { type: string }
        password: { type: string }
    InvitePreview:
      type: object
      required: [email, name, companyName]
      properties:
        email: { type: string }
        name: { type: string }
        companyName: { type: string }
...
    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name: { type: string }
        email: { type: string }
        password: { type: string }
        role: { type: string, enum: [REFERRER] }
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
...